Move `SystemPlugin` upsert logic to a use case for reuse in `SystemStatusStore` and `POSSystemStatusService`.
